onlyCalls/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.1
dataclasses==0.6
decorator==5.1.1
deprecation==2.1.0
elevenlabs==0.2.26
exceptiongroup==1.1.3
executing==2.0.0
frozenlist==1.4.0
gotrue==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
iniconfig==2.0.0
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
multidict==6.0.4
openai==0.28.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.3.0
postgrest==0.11.0
postgrest-py==0.4.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
realtime-py==0.1.3
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
toml==0.10.2
tqdm==4.66.1
traitlets==5.12.0
typing_extensions==4.8.0
urllib3==1.26.18
wcwidth==0.2.8
websockets==10.4
yarl==1.9.2
